Suppress compiler warning mostly -- but -1 fix for uint16_t [skip actions]

+++ main_program/embedded_archive.cpp
@@ -92,7 +92,7 @@ string ArchiveGetFile(std::string_view const arg_filename, std::string &extensio
                 // If prefix_only is true, we only check the beginning of the filename
                 if ( false == archive_filename.starts_with(arg_filename) ) continue;
                 size_t const dot_pos = archive_filename.find('.');
-                if ( (-1 != dot_pos) && ( (dot_pos + 1u) < archive_filename.size() ) )
+                if ( ((size_t)-1 != dot_pos) && ( (dot_pos + 1u) < archive_filename.size() ) )
                 {
                     extension = std::string( archive_filename.substr(dot_pos + 1u) );
                 }
